conn/raw: fix truncation of write() return value
write() returns ssize_t, but the result was stored in an int, which
could truncate on large writes. Use ssize_t for the local variable.

Coverity: CID 392751

+++ conn/raw.c
@@ -324,7 +324,7 @@ int raw_socket_read(struct Connection *conn, char *buf, size_t count)
  */
 int raw_socket_write(struct Connection *conn, const char *buf, size_t count)
 {
-  int rc;
+  ssize_t rc;
   size_t sent = 0;
 
   mutt_sig_allow_interrupt(true);
